# src/routes/taskRoutes.js
// const express = require("express");
// const TaskController = require("../Controllers/taskController");

// const router = express.Router();
// const taskController = new TaskController();

// /**
//  * @swagger
//  * /tasks:
//  *   post:
//  *     summary: Create a new task
//  *     tags: [Tasks]
//  *     requestBody:
//  *       required: true
//  *       content:
//  *         application/json:
//  *           schema:
//  *             type: object
//  *             properties:
//  *               title:
//  *                 type: string
//  *                 example: "New Task"
//  *     responses:
//  *       201:
//  *         description: Task created successfully
//  *       400:
//  *         description: Bad request
//  */
// router.post("/tasks", taskController.createTask.bind(taskController));

// /**
//  * @swagger
//  * /tasks:
//  *   get:
//  *     summary: Get all tasks
//  *     tags: [Tasks]
//  *     responses:
//  *       200:
//  *         description: List of tasks
//  *         content:
//  *           application/json:
//  *             schema:
//  *               type: array
//  *               items:
//  *                 type: object
//  *                 properties:
//  *                   id:
//  *                     type: string
//  *                   title:
//  *                     type: string
//  *                   completed:
//  *                     type: boolean
//  *       500:
//  *         description: Internal server error
//  */
// router.get("/tasks", taskController.getAllTasks.bind(taskController));

// /**
//  * @swagger
//  * /tasks/{id}:
//  *   put:
//  *     summary: Update a task
//  *     tags: [Tasks]
//  *     parameters:
//  *       - in: path
//  *         name: id
//  *         required: true
//  *         schema:
//  *           type: string
//  *         description: Task ID
//  *     requestBody:
//  *       required: true
//  *       content:
//  *         application/json:
//  *           schema:
//  *             type: object
//  *             properties:
//  *               title:
//  *                 type: string
//  *                 example: "Updated Task"
//  *               completed:
//  *                 type: boolean
//  *                 example: true
//  *     responses:
//  *       200:
//  *         description: Task updated successfully
//  *       400:
//  *         description: Bad request
//  *       404:
//  *         description: Task not found
//  */
// router.put("/tasks/:id", taskController.updateTask.bind(taskController));

// /**
//  * @swagger
//  * /tasks/{id}:
//  *   delete:
//  *     summary: Delete a task
//  *     tags: [Tasks]
//  *     parameters:
//  *       - in: path
//  *         name: id
//  *         required: true
//  *         schema:
//  *           type: string
//  *         description: Task ID
//  *     responses:
//  *       200:
//  *         description: Task deleted successfully
//  *       404:
//  *         description: Task not found
//  */
// router.delete("/tasks/:id", taskController.deleteTask.bind(taskController));

// module.exports = router;

const express = require("express");
const TaskController = require("../Controllers/taskController");

const router = express.Router();
const taskController = new TaskController();

/**
 * @swagger
 * /tasks:
 *   post:
 *     summary: Create a new task
 *     tags: [Tasks]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               title:
 *                 type: string
 *                 example: "New Task"
 *               completed:
 *                 type: boolean
 *                 example: false
 *     responses:
 *       201:
 *         description: Task created successfully
 *       400:
 *         description: Bad request
 */
router.post("/tasks", taskController.createTask.bind(taskController));

/**
 * @swagger
 * /tasks:
 *   get:
 *     summary: Get all tasks
 *     tags: [Tasks]
 *     responses:
 *       200:
 *         description: List of tasks
 *         content:
 *           application/json:
 *             schema:
 *               type: array
 *               items:
 *                 type: object
 *                 properties:
 *                   id:
 *                     type: integer
 *                   title:
 *                     type: string
 *                   completed:
 *                     type: boolean
 *       500:
 *         description: Internal server error
 */
router.get("/tasks", taskController.getAllTasks.bind(taskController));

/**
 * @swagger
 * /tasks/{id}:
 *   put:
 *     summary: Update a task
 *     tags: [Tasks]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: integer
 *         description: Task ID
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               title:
 *                 type: string
 *                 example: "Updated Task"
 *               completed:
 *                 type: boolean
 *                 example: true
 *     responses:
 *       200:
 *         description: Task updated successfully
 *       400:
 *         description: Bad request
 *       404:
 *         description: Task not found
 */
router.put("/tasks/:id", taskController.updateTask.bind(taskController));

/**
 * @swagger
 * /tasks/{id}:
 *   delete:
 *     summary: Delete a task
 *     tags: [Tasks]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: integer
 *         description: Task ID
 *     responses:
 *       200:
 *         description: Task deleted successfully
 *       404:
 *         description: Task not found
 */
router.delete("/tasks/:id", taskController.deleteTask.bind(taskController));

module.exports = router;