# swagger.yaml
openapi: 3.0.0
info:
  title: Task Manager API (Frontend Simulated)
  description: A simple task management API simulated in the frontend without a backend.
  version: 1.0.0
paths:
  /tasks:
    post:
      summary: Create a new task
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Nom de la tâche"
                completed:
                  type: boolean
                  example: false
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  title:
                    type: string
                  completed:
                    type: boolean
        '400':
          description: Bad request
    get:
      summary: Get all tasks
      tags:
        - Tasks
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    title:
                      type: string
                    completed:
                      type: boolean
        '500':
          description: Internal server error
  /tasks/{id}:
    put:
      summary: Update a task
      tags:
        - Tasks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Nouveau titre"
                completed:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  title:
                    type: string
                  completed:
                    type: boolean
        '400':
          description: Bad request
        '404':
          description: Task not found
    delete:
      summary: Delete a task
      tags:
        - Tasks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Task ID
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Task deleted successfully"
        '404':
          description: Task not found